[nrf noup] Zeroize the Mbedtls entropy mutex
This makes sure that the content of the mutex
inside the mbedtls_entropy_context is zeroed.
This is a workaround because the CryptoCell
runtime library will generate a fault if
the mutex is not zeroed. This workaround
will be reverted later when NCSDK-17004
is fixed. There is no reason to upstream this
since it is a limitation in our CryptoCell
runtime library and not an upstream limitation.

Ref: NCSDK-8075

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
(cherry picked from commit 2bf6a1f10c4239f088a0049d101651ec9fb7b29f)
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>
(cherry picked from commit 73337db4dada15c23ded33252513c0d67ebf4ed7)
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>