[nrf noup] Constant size for psa_core_key_attributes_t struct
Before the size would change depending on if MBEDTLS_PSA_CRYPTO_KEY_ID_ENCODES_OWNER is defined
or not. This can cause problems when used in different domains. Example:
MBEDTLS_PSA_CRYPTO_KEY_ID_ENCODES_OWNER can not be enabled if PSA_CRYPTO is enabled.
So if building an application as non-secure that uses PSA_CRYPTO
MBEDTLS_PSA_CRYPTO_KEY_ID_ENCODES_OWNER will not be defined. causing id to just be
psa_key_id_t(4 bytes), but TF-M running on the secure domain needs
MBEDTLS_PSA_CRYPTO_KEY_ID_ENCODES_OWNER to be enabled, causing the id field to be a struct
that is 8 byte when TF-M is built it is linked to the cc3xx runtime library that is built with
MBEDTLS_PSA_CRYPTO_KEY_ID_ENCODES_OWNER not defined. As a result, psa_core_key_attributes_t is
interpreted differently within the secure domain.

To just build the provided cc3xx library with MBEDTLS_PSA_CRYPTO_KEY_ID_ENCODES_OWNER enabled
is not a real option as the library is meant to be also used for a secure domain-only application,
that uses PSA_CRYPTO.

This is not going upstream as this only causes a problem if you build some libraries e.g. the
cryptocell not together with mbedtls, but linking them in from a separate build. As this
scenario is not common for upstream we keep this fix downstream.

Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>