[nrf fromtree] Remove redundant component in all.sh
CID is now enabled in the default config (as well as full), so it's
already tested in numerous all.sh components, not need to add one for
that.

We need a component for the legacy/compat option though as it's never
enabled in existing components. So, keep that one, but adjust the name
and fix a typo in a message.

Signed-off-by: Manuel Pégourié-Gonnard <manuel.pegourie-gonnard@arm.com>
(cherry picked from commit 27ba3989d6dbbfcd90f441eec93e7f4788efa8f9)
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>