[nrf noup] Fix buffer overflow in TLS 1.2 ClientKeyExchange parsing
Recreated from: commit c8df898204584e92b9e5001847788d899b34353d
With the following information:

"Fix a buffer overflow in TLS 1.2 ClientKeyExchange parsing. When
MBEDTLS_USE_PSA_CRYPTO is enabled, the length of the public key in an ECDH
or ECDHE key exchange was not validated. This could result in an overflow
of handshake->xxdh_psa_peerkey, overwriting further data in the handshake
structure or further on the heap."

This commit is "noup" since Mbed TLS TLS/DTLS has gone through refactoring
meaning its content had to be recreated.

Signed-off-by: Frank Audun Kvamtrø <frank.kvamtro@nordicsemi.no>