[nrf noup] Fix buffer overflow in TLS 1.3 ECDH public key parsing
Recreated from commit 12c5aaae574114cf4883d0e21aa0045b0a606110
which provides the following information:

"Fix a buffer overflow in TLS 1.3 ServerHello and ClientHello parsing. The
length of the public key in an ECDH- or FFDH-based key exchange was not
validated. This could result in an overflow of handshake->xxdh_psa_peerkey,
overwriting further data in the handshake structure or further on the
heap."

This commit is "noup" since it TLS/DTLS is undergoing refactoring and
the content of the commit had to be recreated.

Signed-off-by: Frank Audun Kvamtrø <frank.kvamtro@nordicsemi.no>