[nrf noup]  Fix buffer overread with stream cipher
Recreated from commit faf0b8604ac49456b0cff7a34ad27485ca145cce
which provides the following information

"With stream ciphers, add a check that there's enough room to read a MAC
in the record. Without this check, subtracting the MAC length from the
data length resulted in an integer underflow, causing the MAC calculation
to try reading (SIZE_MAX + 1 - maclen) bytes of input, which is a buffer
overread."

This commit is a "noup" since TLS/DTLS is undergoing refactoring and
the content of the commit had to be recreated.

Signed-off-by: Frank Audun Kvamtrø <frank.kvamtro@nordicsemi.no>