[nrf noup] Make the md contexts fixed size
The CryptoCell runtime libraries are using the
mbedtls_md APIs for hashing. Since the CryptoCell
libraries are binaries and the mbedtls_md APIs are
compiled from source at build time, they need to have
the same view of the mbedtls_md context.

This change makes the mbedtls_md contexts to have
fixed sizes.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>