[nrf noup] Move headers also in Oberon PSA core
The Oberon PSA core has some replacement headers but it also
uses headers from the the library folder. Since the headers and the
C files in this folder perform includsion with double quotes this
can create issues since the Oberon PSA core may use headrs from
this repo that it shouldn't.

This moves the headers that the Oberon PSA core replaces with
to another folder so that they don't accidentaly included in the
Oberon PSA core.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>